Separate expression handle treatment
